feat: finshed backend password change
